Embeddings parsing, with added pre-trained.
